# lib/api-zod/src/generated/types/contactInput.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface ContactInput {
  /** @minLength 2 */
  name: string;
  email: string;
  /** @minLength 8 */
  phone: string;
  /** @minLength 2 */
  subject: string;
  /** @minLength 10 */
  message: string;
}